move L2projector to feec/mass.py (logically better and avoids circular imports); further adapt spline_attributes
