+++ src/struphy/propagators/tests/test_gyrokinetic_poisson.py
@@ -7,7 +7,7 @@ from feectools.ddm.mpi import mpi as MPI
 
 from struphy import domains
 from struphy.feec.mass import WeightedMassOperators
-from struphy.feec.projectors import L2Projector
+from struphy.feec.mass import L2Projector
 from struphy.feec.psydac_derham import Derham
 from struphy.geometry.base import Domain
 from struphy.io.options import DerhamOptions
