+++ src/struphy/propagators/tests/test_poisson.py
@@ -14,7 +14,7 @@ from struphy import (
     perturbations,
 )
 from struphy.feec.mass import WeightedMassOperators
-from struphy.feec.projectors import L2Projector
+from struphy.feec.mass import L2Projector
 from struphy.feec.psydac_derham import Derham
 from struphy.geometry.base import Domain
 from struphy.io.options import DerhamOptions
